backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
setuptools==75.1.0
sqlparse==0.5.3
uritemplate==4.1.1
wheel==0.44.0
